Remove cache_strategies.py and inline into core_cache.py
- Delete unnecessary cache_strategies.py file
- Move CacheStrategy enum and _get_cache_strategy() into core_cache.py
- Simpler code structure with one less module
